+++ src/lyscripts/compute/__init__.py
@@ -5,16 +5,17 @@ This can in turn speed up the computation of risks and prevalences.
 
 from pydantic_settings import BaseSettings, CliApp, CliSubCommand
 
-from lyscripts.compute import posteriors, prevalences, priors, risks
+from lyscripts.compute import posteriors, prevalences, priors, risks, evidence
 
 
 class ComputeCLI(BaseSettings):
-    """Compute priors, posteriors, risks, and prevalences from model samples."""
+    """Compute priors, posteriors, risks, prevalences and model evidence from model samples."""
 
     priors: CliSubCommand[priors.PriorsCLI]
     posteriors: CliSubCommand[posteriors.PosteriorsCLI]
     risks: CliSubCommand[risks.RisksCLI]
     prevalences: CliSubCommand[prevalences.PrevalencesCLI]
+    evidence: CliSubCommand[evidence.EvidenceCLI]
 
     def cli_cmd(self) -> None:
         """Start the ``compute`` subcommand."""