+++ src/lyscripts/data/enhance.py
@@ -1,13 +1,15 @@
 """Enhance the dataset by inferring additional columns from the data.
 
-This is a command-line interface to the
-:py:func:`~lydata.utils.infer_and_combine_levels` function.
+This is a command-line interface to the methods
+:py:meth:`~lydata.accessor.LyDataAccessor.combine` and
+:py:meth:`~lydata.accessor.LyDataAccessor.augment` of the
+:py:class:`~lydata.accessor.LyDataAccessor` class.
 """
 
 from typing import Literal
 
 from loguru import logger
-from lydata import infer_and_combine_levels
+from lydata.accessor import LyDataFrame
 from lydata.utils import ModalityConfig
 
 from lyscripts.cli import assemble_main
@@ -21,7 +23,6 @@ class EnhanceCLI(BaseCLI):
     input: DataConfig
     modalities: dict[str, ModalityConfig] | None = None
     method: Literal["max_llh", "rank"] = "max_llh"
-    sides: list[Literal["ipsi", "contra"]] = ["ipsi", "contra"]
     lnl_subdivisions: dict[str, list[str]] = {
         "I": ["a", "b"],
         "II": ["a", "b"],
@@ -33,29 +34,18 @@ class EnhanceCLI(BaseCLI):
         """Infer additional columns from the data and save the enhanced dataset.
 
         This basically provides a CLI to the
-        :py:func:`~lydata.utils.infer_and_combine_levels` function. See its docs for
+        :py:func:`~lydata.accessor.LyDataAccessor.augment` function. See its docs for
         more details on what exactly is happening here.
         """
         logger.debug(self.model_dump_json(indent=2))
 
-        data = self.input.load()
-        modality_names = list(self.modalities.keys()) if self.modalities else None
-
-        infer_lvls_kwargs = {
-            "modalities": modality_names,
-            "sides": self.sides,
-            "subdivisions": self.lnl_subdivisions,
-        }
-        enhanced = infer_and_combine_levels(
-            dataset=data,
-            infer_superlevels_kwargs=infer_lvls_kwargs,
-            infer_sublevels_kwargs=infer_lvls_kwargs,
-            combine_kwargs={
-                "modalities": self.modalities,
-                "method": self.method,
-            },
+        data: LyDataFrame = self.input.load()
+        data = data.ly.enhance(
+            modalities=self.modalities,
+            method=self.method,
+            subdivisions=self.lnl_subdivisions,
         )
-        save_table_to_csv(file_path=self.output_file, table=enhanced)
+        save_table_to_csv(file_path=self.output_file, table=data)
 
 
 if __name__ == "__main__":